requirements.txt

astroid==3.1.0
coverage==7.4.3
dill==0.3.8
exceptiongroup==1.2.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pylint==3.1.0
pytest==8.0.2
pytest-cov==4.1.0
pytest-pylint==0.21.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
python-dotenv
pandas
